+++ python/api/CHANGELOG.rst
@@ -1,3 +1,17 @@
+0.2.2 (2016-03-30)
+==================
+* Refresh token bugfixes
+
+0.2.1 (2016-03-28)
+==================
+* Update requirements to use requests-oauthlib>=0.6.1
+
+0.2 (2016-03-23)
+================
+
+* Drop OAuth1 support. See `OAuth1 deprecated <https://dev.fitbit.com/docs/oauth2/#oauth-1-0a-deprecated>`_
+* Drop py26 and py32 support
+
 0.1.3 (2015-02-04)
 ==================